Basic Structure ready for solving
Since I want implement a TDD approach, I've created a directory for solutions and a directory for unit tests. From here, each new problem will consist of the creation of a new feature branch. Each branch will contain (after rebasing) only two commits. The first is a comprehensive set of tests, the second the solution. Any mistakes that I later notice will be placed on a new Feature branch, with the relevant fix